[refactor] flipping configuration file generation strategy.
Flips the code generation around for configuration and manifest to be
defined in dhall files. GHC then evaluates a dhall type expression at compile
time. This allows for manual evolution of the Dhall code, which means easier
documentation, code factoring, code presentation, and reduction in amount
of generated files, so less duplication. Examples are also pulled out of the
Haskell source and refactored to be more presentable. Cabal file generation
is also removed.