[fix] modify downstream API to include timestamps
There's an ongoing issue with the monitoring API due to the fact that
the timestamps for events are attached when the daemon receives them
instead of when the measure was made. It can trigger abnormal values to
be computed by NRM if monitoring messages are delayed or occur in
batches.

This commit changes the downstream API format so that our perfwrapper
can report its own timing. Similar modifications will be made in the
future for other events.