[down] add general timestamps to downstream
Add a timestamp typed int64 to the downstream API, to ensure that
whoever is in charge of reporting performance information properly
reports the time of the event, instead of waiting for the message to be
processed.

This has significant impact on the API because of the resulting change
in format across haskell, python, and C code. In particular, the message
schema for downstream gains an extra tuple at the top, that complicates
a bit the current logic for packing/unpacking.